+++ great_expectations/dataset/dataset.py
@@ -342,7 +342,7 @@ class Dataset(MetaDataset):
             **kwargs       : Keyword arguments to be passed the the function
 
         Returns:
-            A JSON-serializable expectation result object.
+            An ExpectationSuiteValidationResult
 
         Notes:
             This function is a thin layer to allow quick testing of new expectation functions, without having to \
@@ -364,7 +364,7 @@ class Dataset(MetaDataset):
             **kwargs       : Keyword arguments to be passed the the function
 
         Returns:
-            A JSON-serializable expectation result object.
+            An ExpectationSuiteValidationResult
 
         Notes:
             This function is a thin layer to allow quick testing of new expectation functions, without having to \
@@ -422,7 +422,7 @@ class Dataset(MetaDataset):
                 For more detail, see :ref:`meta`.
 
         Returns:
-            A JSON-serializable expectation result object.
+            An ExpectationSuiteValidationResult
 
             Exact fields vary depending on the values passed to :ref:`result_format <result_format>` and
             :ref:`include_config`, :ref:`catch_exceptions`, and :ref:`meta`.
@@ -473,7 +473,7 @@ class Dataset(MetaDataset):
                 modification. For more detail, see :ref:`meta`.
 
         Returns:
-            A JSON-serializable expectation result object.
+            An ExpectationSuiteValidationResult
 
             Exact fields vary depending on the values passed to :ref:`result_format <result_format>` and
             :ref:`include_config`, :ref:`catch_exceptions`, and :ref:`meta`.
@@ -544,7 +544,7 @@ class Dataset(MetaDataset):
                 modification. For more detail, see :ref:`meta`.
 
         Returns:
-            A JSON-serializable expectation result object.
+            An ExpectationSuiteValidationResult
 
             Exact fields vary depending on the values passed to :ref:`result_format <result_format>` and
             :ref:`include_config`, :ref:`catch_exceptions`, and :ref:`meta`.
@@ -625,7 +625,7 @@ class Dataset(MetaDataset):
                 modification. For more detail, see :ref:`meta`.
 
         Returns:
-            A JSON-serializable expectation result object.
+            An ExpectationSuiteValidationResult
 
             Exact fields vary depending on the values passed to :ref:`result_format <result_format>` and
             :ref:`include_config`, :ref:`catch_exceptions`, and :ref:`meta`.
@@ -685,7 +685,7 @@ class Dataset(MetaDataset):
                 modification. For more detail, see :ref:`meta`.
 
         Returns:
-            A JSON-serializable expectation result object.
+            An ExpectationSuiteValidationResult
 
             Exact fields vary depending on the values passed to :ref:`result_format <result_format>` and
             :ref:`include_config`, :ref:`catch_exceptions`, and :ref:`meta`.
@@ -766,7 +766,7 @@ class Dataset(MetaDataset):
                 modification. For more detail, see :ref:`meta`.
 
         Returns:
-            A JSON-serializable expectation result object.
+            An ExpectationSuiteValidationResult
 
             Exact fields vary depending on the values passed to :ref:`result_format <result_format>` and
             :ref:`include_config`, :ref:`catch_exceptions`, and :ref:`meta`.
@@ -833,7 +833,7 @@ class Dataset(MetaDataset):
                 modification. For more detail, see :ref:`meta`.
 
         Returns:
-            A JSON-serializable expectation result object.
+            An ExpectationSuiteValidationResult
 
             Exact fields vary depending on the values passed to :ref:`result_format <result_format>` and
             :ref:`include_config`, :ref:`catch_exceptions`, and :ref:`meta`.
@@ -881,7 +881,7 @@ class Dataset(MetaDataset):
                 modification. For more detail, see :ref:`meta`.
 
         Returns:
-            A JSON-serializable expectation result object.
+            An ExpectationSuiteValidationResult
 
             Exact fields vary depending on the values passed to :ref:`result_format <result_format>` and
             :ref:`include_config`, :ref:`catch_exceptions`, and :ref:`meta`.
@@ -931,7 +931,7 @@ class Dataset(MetaDataset):
                 modification. For more detail, see :ref:`meta`.
 
         Returns:
-            A JSON-serializable expectation result object.
+            An ExpectationSuiteValidationResult
 
             Exact fields vary depending on the values passed to :ref:`result_format <result_format>` and
             :ref:`include_config`, :ref:`catch_exceptions`, and :ref:`meta`.
@@ -996,7 +996,7 @@ class Dataset(MetaDataset):
                 modification. For more detail, see :ref:`meta`.
 
         Returns:
-            A JSON-serializable expectation result object.
+            An ExpectationSuiteValidationResult
 
             Exact fields vary depending on the values passed to :ref:`result_format <result_format>` and
             :ref:`include_config`, :ref:`catch_exceptions`, and :ref:`meta`.
@@ -1058,7 +1058,7 @@ class Dataset(MetaDataset):
                 modification. For more detail, see :ref:`meta`.
 
         Returns:
-            A JSON-serializable expectation result object.
+            An ExpectationSuiteValidationResult
 
             Exact fields vary depending on the values passed to :ref:`result_format <result_format>` and
             :ref:`include_config`, :ref:`catch_exceptions`, and :ref:`meta`.
@@ -1141,7 +1141,7 @@ class Dataset(MetaDataset):
                 modification. For more detail, see :ref:`meta`.
 
         Returns:
-            A JSON-serializable expectation result object.
+            An ExpectationSuiteValidationResult
 
             Exact fields vary depending on the values passed to :ref:`result_format <result_format>` and
             :ref:`include_config`, :ref:`catch_exceptions`, and :ref:`meta`.
@@ -1216,7 +1216,7 @@ class Dataset(MetaDataset):
                 modification. For more detail, see :ref:`meta`.
 
         Returns:
-            A JSON-serializable expectation result object.
+            An ExpectationSuiteValidationResult
 
             Exact fields vary depending on the values passed to :ref:`result_format <result_format>` and
             :ref:`include_config`, :ref:`catch_exceptions`, and :ref:`meta`.
@@ -1287,7 +1287,7 @@ class Dataset(MetaDataset):
                 modification. For more detail, see :ref:`meta`.
 
         Returns:
-            A JSON-serializable expectation result object.
+            An ExpectationSuiteValidationResult
 
             Exact fields vary depending on the values passed to :ref:`result_format <result_format>` and
             :ref:`include_config`, :ref:`catch_exceptions`, and :ref:`meta`.
@@ -1354,7 +1354,7 @@ class Dataset(MetaDataset):
                 modification. For more detail, see :ref:`meta`.
 
         Returns:
-            A JSON-serializable expectation result object.
+            An ExpectationSuiteValidationResult
 
             Exact fields vary depending on the values passed to :ref:`result_format <result_format>` and
             :ref:`include_config`, :ref:`catch_exceptions`, and :ref:`meta`.
@@ -1416,7 +1416,7 @@ class Dataset(MetaDataset):
                 modification. For more detail, see :ref:`meta`.
 
         Returns:
-            A JSON-serializable expectation result object.
+            An ExpectationSuiteValidationResult
 
             Exact fields vary depending on the values passed to :ref:`result_format <result_format>` and
             :ref:`include_config`, :ref:`catch_exceptions`, and :ref:`meta`.
@@ -1480,7 +1480,7 @@ class Dataset(MetaDataset):
                 modification. For more detail, see :ref:`meta`.
 
         Returns:
-            A JSON-serializable expectation result object.
+            An ExpectationSuiteValidationResult
 
             Exact fields vary depending on the values passed to :ref:`result_format <result_format>` and
             :ref:`include_config`, :ref:`catch_exceptions`, and :ref:`meta`.
@@ -1542,7 +1542,7 @@ class Dataset(MetaDataset):
                 modification. For more detail, see :ref:`meta`.
 
         Returns:
-            A JSON-serializable expectation result object.
+            An ExpectationSuiteValidationResult
 
             Exact fields vary depending on the values passed to :ref:`result_format <result_format>` and
             :ref:`include_config`, :ref:`catch_exceptions`, and :ref:`meta`.
@@ -1597,7 +1597,7 @@ class Dataset(MetaDataset):
                 modification. For more detail, see :ref:`meta`.
 
         Returns:
-            A JSON-serializable expectation result object.
+            An ExpectationSuiteValidationResult
 
             Exact fields vary depending on the values passed to :ref:`result_format <result_format>` and
             :ref:`include_config`, :ref:`catch_exceptions`, and :ref:`meta`.
@@ -1655,7 +1655,7 @@ class Dataset(MetaDataset):
                 modification. For more detail, see :ref:`meta`.
 
         Returns:
-            A JSON-serializable expectation result object.
+            An ExpectationSuiteValidationResult
 
             Exact fields vary depending on the values passed to :ref:`result_format <result_format>` and
             :ref:`include_config`, :ref:`catch_exceptions`, and :ref:`meta`.
@@ -1717,7 +1717,7 @@ class Dataset(MetaDataset):
                 modification. For more detail, see :ref:`meta`.
 
         Returns:
-            A JSON-serializable expectation result object.
+            An ExpectationSuiteValidationResult
 
             Exact fields vary depending on the values passed to :ref:`result_format <result_format>` and
             :ref:`include_config`, :ref:`catch_exceptions`, and :ref:`meta`.
@@ -1774,7 +1774,7 @@ class Dataset(MetaDataset):
                 modification. For more detail, see :ref:`meta`.
 
         Returns:
-            A JSON-serializable expectation result object.
+            An ExpectationSuiteValidationResult
 
             Exact fields vary depending on the values passed to :ref:`result_format <result_format>` and
             :ref:`include_config`, :ref:`catch_exceptions`, and :ref:`meta`.
@@ -1833,7 +1833,7 @@ class Dataset(MetaDataset):
                 modification. For more detail, see :ref:`meta`.
 
         Returns:
-            A JSON-serializable expectation result object.
+            An ExpectationSuiteValidationResult
 
             Exact fields vary depending on the values passed to :ref:`result_format <result_format>` and
             :ref:`include_config`, :ref:`catch_exceptions`, and :ref:`meta`.
@@ -1879,7 +1879,7 @@ class Dataset(MetaDataset):
                 modification. For more detail, see :ref:`meta`.
 
         Returns:
-            A JSON-serializable expectation result object.
+            An ExpectationSuiteValidationResult
 
             Exact fields vary depending on the values passed to :ref:`result_format <result_format>` and
             :ref:`include_config`, :ref:`catch_exceptions`, and :ref:`meta`.
@@ -1925,7 +1925,7 @@ class Dataset(MetaDataset):
                 modification. For more detail, see :ref:`meta`.
 
         Returns:
-            A JSON-serializable expectation result object.
+            An ExpectationSuiteValidationResult
 
             Exact fields vary depending on the values passed to :ref:`result_format <result_format>` and
             :ref:`include_config`, :ref:`catch_exceptions`, and :ref:`meta`.
@@ -1976,7 +1976,7 @@ class Dataset(MetaDataset):
                 modification. For more detail, see :ref:`meta`.
 
         Returns:
-            A JSON-serializable expectation result object.
+            An ExpectationSuiteValidationResult
 
             Exact fields vary depending on the values passed to :ref:`result_format <result_format>` and
             :ref:`include_config`, :ref:`catch_exceptions`, and :ref:`meta`.
@@ -2050,7 +2050,7 @@ class Dataset(MetaDataset):
                 modification. For more detail, see :ref:`meta`.
 
         Returns:
-            A JSON-serializable expectation result object.
+            An ExpectationSuiteValidationResult
 
             Exact fields vary depending on the values passed to :ref:`result_format <result_format>` and
             :ref:`include_config`, :ref:`catch_exceptions`, and :ref:`meta`.
@@ -2156,7 +2156,7 @@ class Dataset(MetaDataset):
                 modification. For more detail, see :ref:`meta`.
 
         Returns:
-            A JSON-serializable expectation result object.
+            An ExpectationSuiteValidationResult
 
             Exact fields vary depending on the values passed to :ref:`result_format <result_format>` and
             :ref:`include_config`, :ref:`catch_exceptions`, and :ref:`meta`.
@@ -2257,7 +2257,7 @@ class Dataset(MetaDataset):
                 modification. For more detail, see :ref:`meta`.
 
         Returns:
-            A JSON-serializable expectation result object.
+            An ExpectationSuiteValidationResult
 
             Exact fields vary depending on the values passed to :ref:`result_format <result_format>` and
             :ref:`include_config`, :ref:`catch_exceptions`, and :ref:`meta`.
@@ -2347,7 +2347,7 @@ class Dataset(MetaDataset):
                 modification. For more detail, see :ref:`meta`.
 
         Returns:
-            A JSON-serializable expectation result object.
+            An ExpectationSuiteValidationResult
 
             Exact fields vary depending on the values passed to :ref:`result_format <result_format>` and
             :ref:`include_config`, :ref:`catch_exceptions`, and :ref:`meta`.
@@ -2422,7 +2422,7 @@ class Dataset(MetaDataset):
                 modification. For more detail, see :ref:`meta`.
 
         Returns:
-            A JSON-serializable expectation result object.
+            An ExpectationSuiteValidationResult
 
             Exact fields vary depending on the values passed to :ref:`result_format <result_format>` and
             :ref:`include_config`, :ref:`catch_exceptions`, and :ref:`meta`.
@@ -2533,7 +2533,7 @@ class Dataset(MetaDataset):
                 modification. For more detail, see :ref:`meta`.
 
         Returns:
-            A JSON-serializable expectation result object.
+            An ExpectationSuiteValidationResult
 
             Exact fields vary depending on the values passed to :ref:`result_format <result_format>` and
             :ref:`include_config`, :ref:`catch_exceptions`, and :ref:`meta`.
@@ -2677,7 +2677,7 @@ class Dataset(MetaDataset):
                 modification. For more detail, see :ref:`meta`.
 
         Returns:
-            A JSON-serializable expectation result object.
+            An ExpectationSuiteValidationResult
 
             Exact fields vary depending on the values passed to :ref:`result_format <result_format>` and
             :ref:`include_config`, :ref:`catch_exceptions`, and :ref:`meta`.
@@ -2777,7 +2777,7 @@ class Dataset(MetaDataset):
                 modification. For more detail, see :ref:`meta`.
 
         Returns:
-            A JSON-serializable expectation result object.
+            An ExpectationSuiteValidationResult
 
             Exact fields vary depending on the values passed to :ref:`result_format <result_format>` and
             :ref:`include_config`, :ref:`catch_exceptions`, and :ref:`meta`.
@@ -2871,7 +2871,7 @@ class Dataset(MetaDataset):
                 modification. For more detail, see :ref:`meta`.
 
         Returns:
-            A JSON-serializable expectation result object.
+            An ExpectationSuiteValidationResult
 
             Exact fields vary depending on the values passed to :ref:`result_format <result_format>` and
             :ref:`include_config`, :ref:`catch_exceptions`, and :ref:`meta`.
@@ -2963,7 +2963,7 @@ class Dataset(MetaDataset):
                 modification. For more detail, see :ref:`meta`.
 
         Returns:
-            A JSON-serializable expectation result object.
+            An ExpectationSuiteValidationResult
 
             Exact fields vary depending on the values passed to :ref:`result_format <result_format>` and
             :ref:`include_config`, :ref:`catch_exceptions`, and :ref:`meta`.
@@ -3069,7 +3069,7 @@ class Dataset(MetaDataset):
                 modification. For more detail, see :ref:`meta`.
 
         Returns:
-            A JSON-serializable expectation result object.
+            An ExpectationSuiteValidationResult
 
             Exact fields vary depending on the values passed to :ref:`result_format <result_format>` and
             :ref:`include_config`, :ref:`catch_exceptions`, and :ref:`meta`.
@@ -3144,7 +3144,7 @@ class Dataset(MetaDataset):
                 modification. For more detail, see :ref:`meta`.
 
         Returns:
-            A JSON-serializable expectation result object.
+            An ExpectationSuiteValidationResult
 
             Exact fields vary depending on the values passed to :ref:`result_format <result_format>` and
             :ref:`include_config`, :ref:`catch_exceptions`, and :ref:`meta`.
@@ -3251,7 +3251,7 @@ class Dataset(MetaDataset):
                 modification. For more detail, see :ref:`meta`.
 
         Returns:
-            A JSON-serializable expectation result object.
+            An ExpectationSuiteValidationResult
 
             Exact fields vary depending on the values passed to :ref:`result_format <result_format>` and
             :ref:`include_config`, :ref:`catch_exceptions`, and :ref:`meta`.
@@ -3388,7 +3388,7 @@ class Dataset(MetaDataset):
                 modification. For more detail, see :ref:`meta`.
 
         Returns:
-            A JSON-serializable expectation result object.
+            An ExpectationSuiteValidationResult
 
             Exact fields vary depending on the values passed to :ref:`result_format <result_format>` and
             :ref:`include_config`, :ref:`catch_exceptions`, and :ref:`meta`.
@@ -3525,7 +3525,7 @@ class Dataset(MetaDataset):
                 modification. For more detail, see :ref:`meta`.
 
         Returns:
-            A JSON-serializable expectation result object.
+            An ExpectationSuiteValidationResult
 
             Exact fields vary depending on the values passed to :ref:`result_format <result_format>` and
             :ref:`include_config`, :ref:`catch_exceptions`, and :ref:`meta`.
@@ -3656,7 +3656,7 @@ class Dataset(MetaDataset):
                 modification. For more detail, see :ref:`meta`.
 
         Returns:
-            A JSON-serializable expectation result object.
+            An ExpectationSuiteValidationResult
 
             Exact fields vary depending on the values passed to :ref:`result_format <result_format>` and
             :ref:`include_config`, :ref:`catch_exceptions`, and :ref:`meta`.
@@ -3763,7 +3763,7 @@ class Dataset(MetaDataset):
                 modification. For more detail, see :ref:`meta`.
 
         Returns:
-            A JSON-serializable expectation result object.
+            An ExpectationSuiteValidationResult
 
             Exact fields vary depending on the values passed to :ref:`result_format <result_format>` and
             :ref:`include_config`, :ref:`catch_exceptions`, and :ref:`meta`.
@@ -4182,7 +4182,7 @@ class Dataset(MetaDataset):
                 modification. For more detail, see :ref:`meta`.
 
         Returns:
-            A JSON-serializable expectation result object.
+            An ExpectationSuiteValidationResult
 
             Exact fields vary depending on the values passed to :ref:`result_format <result_format>` and
             :ref:`include_config`, :ref:`catch_exceptions`, and :ref:`meta`.
@@ -4233,7 +4233,7 @@ class Dataset(MetaDataset):
                 modification. For more detail, see :ref:`meta`.
 
         Returns:
-            A JSON-serializable expectation result object.
+            An ExpectationSuiteValidationResult
 
             Exact fields vary depending on the values passed to :ref:`result_format <result_format>` and
             :ref:`include_config`, :ref:`catch_exceptions`, and :ref:`meta`.
@@ -4278,7 +4278,7 @@ class Dataset(MetaDataset):
                 modification. For more detail, see :ref:`meta`.
 
         Returns:
-            A JSON-serializable expectation result object.
+            An ExpectationSuiteValidationResult
 
             Exact fields vary depending on the values passed to :ref:`result_format <result_format>` and
             :ref:`include_config`, :ref:`catch_exceptions`, and :ref:`meta`.
@@ -4325,7 +4325,7 @@ class Dataset(MetaDataset):
                 modification. For more detail, see :ref:`meta`.
 
         Returns:
-            A JSON-serializable expectation result object.
+            An ExpectationSuiteValidationResult
 
             Exact fields vary depending on the values passed to :ref:`result_format <result_format>` and
             :ref:`include_config`, :ref:`catch_exceptions`, and :ref:`meta`.