+++ docs/reference/changelog.rst
@@ -6,6 +6,8 @@ Changelog
 
 Develop
 -----------------
+* [FEATURE] Add new expectation expect_table_row_count_to_equal_other_table for
+ SqlAlchemyDataset
 * [ENHANCEMENT] Added 3 new usage stats events: "cli.new_ds_choice", "data_context.add_datasource", and "datasource.sqlalchemy.connect"
 * [FEATURE] Add support for expect_column_values_to_match_regex_list exception for Spark backend
 * [BUGFIX] fixed issue where calling head() on a SqlAlchemyDataset would fail if the underlying table is empty